reorg protocol
