+++ beep/protocol/__init__.py
@@ -1,0 +1,8 @@
+#  Copyright (c) 2019 Toyota Research Institute
+import os
+
+
+MODULE_DIR = os.path.dirname(__file__)
+PROCEDURE_TEMPLATE_DIR = os.path.join(MODULE_DIR, "procedure_templates")
+SCHEDULE_TEMPLATE_DIR = os.path.join(MODULE_DIR, "schedule_templates")
+PROTOCOL_SCHEMA_DIR = os.path.join(MODULE_DIR, "protocol_schemas")
