+++ beep/protocol/procedure_templates/diagnosticV1.000
