+++ beep/protocol/procedure_templates/diagnosticV2.000
