+++ beep/protocol/procedure_templates/diagnosticV3.000
