+++ beep/protocol/procedure_templates/diagnosticV4.000
