+++ beep/protocol/protocol_schemas/arbin_schedule_schema.yaml
