+++ beep/protocol/schedule_templates/20170630-3_6C_9per_5C.sdu
